refact: move Input to atoms
